New component schema
I think it'll be more logical in the long run to split the objects
created based on specifications in the JSON files from what we're
fetching from the package repositories. The repositories are
authoritative and humans have a knack for mistyping things.

We definitely need a better way to hook package sources in when
resolving versions though -- at the moment we're reliant upon
repositories providing all the sources.