Fix installation with Composer
This broke because we were assuming that the ```vendor``` directory
would always be located within our directory, but this is only the case
for in-place installations where a ```composer install``` is executed
from our directory.

Composer instead bundles all of our dependencies into a shared
```vendor``` directory of which we're a subdirectory. This fix applies
to both local and global installations.

Closes #16